Move voting components into a new folder
