+++ packages/lesswrong/components/votes/VoteButton.jsx
