+++ packages/lesswrong/lib/components.js
@@ -37,10 +37,8 @@ import '../components/common/Section.jsx';
 import '../components/common/SectionSubtitle.jsx';
 import '../components/common/MetaInfo.jsx';
 import '../components/common/NoContent.jsx';
-import '../components/common/VoteButton.jsx';
 import '../components/common/SearchBar.jsx';
 import '../components/common/DialogGroup.jsx';
-import '../components/common/VotesInfo.jsx';
 import '../components/common/DraftJSRenderer.jsx';
 import '../components/common/HoverOver.jsx';
 import '../components/common/ErrorBoundary.jsx';
@@ -111,11 +109,15 @@ import '../components/posts/PostsEditPage.jsx';
 import '../components/posts/PostsListHeader.jsx';
 import '../components/posts/FeaturedPostsPage.jsx';
 import '../components/posts/PostsGroupDetails.jsx';
-import '../components/posts/PostsVote.jsx';
 import '../components/posts/PostsStats.jsx';
 import '../components/posts/TableOfContents';
 import '../components/posts/ShowOrHideHighlightButton.jsx';
 import '../components/posts/PostsUserAndCoauthors.jsx';
+
+import '../components/votes/VoteButton.jsx';
+import '../components/votes/VotesInfo.jsx';
+import '../components/comments/CommentsVote.jsx';
+import '../components/posts/PostsVote.jsx';
 
 import '../components/localGroups/CommunityHome.jsx';
 import '../components/localGroups/CommunityMap.jsx';
@@ -158,7 +160,6 @@ import '../components/comments/CommentsList.jsx';
 import '../components/comments/CommentsLoadMore.jsx';
 import '../components/comments/CommentsNode.jsx';
 import '../components/comments/CommentsViews.jsx';
-import '../components/comments/CommentsVote.jsx';
 import '../components/comments/RecentComments.jsx';
 import '../components/comments/RecentCommentsItem.jsx';
 import '../components/comments/RecentCommentsSingle.jsx';
